Simplify MaNGA grouped HDF5 reader path
